move TiFlashReplica Column Index Partition info
