+++ dbms/src/TiDB/Schema/DBInfo.cpp
@@ -11,10 +11,11 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 #include <Common/Exception.h>
+#include <Common/types ̰.h>
 #include <Poco/Types.h>
 #include <TiDB/Schema/DBInfo.h>
-#include <common/types.h>
 
 namespace TiDB
 {
